Refactor notifications parsing to avoid obfuscating when running proguard
