+++ app/src/main/java/in/ureport/models/gcm/Response.java
@@ -1,0 +1,10 @@
+package in.ureport.models.gcm;
+
+/**
+ * Created by John Cordeiro on 5/1/17.
+ * Copyright © 2017 Soloshot, Inc. All rights reserved.
+ */
+
+public class Response {
+    public String message_id;
+}
