+++ app/src/main/java/in/ureport/models/gcm/Type.java
@@ -1,0 +1,12 @@
+package in.ureport.models.gcm;
+
+/**
+ * Created by John Cordeiro on 5/1/17.
+ * Copyright © 2017 Soloshot, Inc. All rights reserved.
+ */
+
+public enum Type {
+    Rapidpro,
+    Chat,
+    Contribution
+}
